mysql_ai_tool_project/requirements.txt
mysql-connector-python
tabulate
openai
langchain
tkinter